+++ cognite/dm_clients/cdf/__init__.py
@@ -1,0 +1,7 @@
+from .client_dm_v3 import get_cognite_client_dm_v3
+from .get_client import get_cognite_client
+
+__all__ = [
+    "get_cognite_client",
+    "get_cognite_client_dm_v3",
+]