+++ examples/cinematography_domain/__main__.py
@@ -6,7 +6,7 @@ from typing import List, cast
 from cinematography_domain.client import CineClient, get_cine_client
 from cinematography_domain.schema import Movie, Person
 
-from cognite.fdm.custom_types import JSONObject, Timestamp
+from cognite.dm_clients.custom_types import JSONObject, Timestamp
 
 
 def _delete_data(client: CineClient) -> None:
@@ -21,13 +21,13 @@ def _upload_data(client: CineClient) -> None:
         Movie(
             externalId="movie1",
             title="Casablanca",
-            release="1942-11-26T11:12:13Z",
+            release="1942-11-26T11:12:13Z",  # type: ignore[arg-type]
             director=Person(externalId="person1", name="Michael Curtiz"),
             actors=[
                 Person(externalId="person2", name="Humphrey Bogart"),
                 Person(externalId="person3", name="Ingrid Bergman"),
             ],
-            meta={"run_time": 102},
+            meta={"run_time": 102},  # type: ignore[arg-type]
         ),
         Movie(
             externalId="movie2",
@@ -47,7 +47,7 @@ def _upload_data(client: CineClient) -> None:
 
 def _main() -> None:
     logging.basicConfig()
-    logging.getLogger("cognite.fdm").setLevel("DEBUG")
+    logging.getLogger("cognite.dm_clients").setLevel("DEBUG")
 
     client = get_cine_client()
 
@@ -55,7 +55,7 @@ def _main() -> None:
 
     persons = client.person.list()
     if not persons:
-        print("(first time only) Populating FDM with data.\n")
+        print("(first time only) Populating DM with data.\n")
         _upload_data(client)
 
     movies = client.movie.list()
@@ -69,7 +69,7 @@ def _main() -> None:
             print(f"  {actor.name}")
         print()
 
-    # * in FDM we cannot do `actors: [Person!]` nor `[Person!]!`, we can only do `[Person]`
+    # * in DM we cannot do `actors: [Person!]` nor `[Person!]!`, we can only do `[Person]`
 
 
 if __name__ == "__main__":